Refactor headers in panda/board/boards/*.h
