+++ SConscript
@@ -117,7 +117,7 @@ def build_project(project_name, project, extra_flags):
   env.Objcopy(f"obj/bootstub.{project_name}.bin", bootstub_elf)
 
   # Collect all C files from the directory
-  board_sources = glob.glob(f"{panda_root}/board/boards/source/*.c")
+  board_sources = glob.glob(f"{panda_root}/board/boards/**/*.c")
 
   # Turn them into object files, namespaced by project
   board_objects = [env.Object(f"obj/{project_name}_{os.path.basename(src).replace('.c', '')}", src) for src in board_sources]
