+++ board/boards/board_declarations.h
@@ -2,6 +2,7 @@
 
 #include <stdint.h>
 #include <stdbool.h>
+#include "../drivers/harness_declarations.h"
 
 // ******************** Prototypes ********************
 typedef enum {
