+++ board/boards/src/board_black.c
@@ -1,6 +1,4 @@
-#pragma once
-
-#include "board_declarations.h"
+#include "../board_declarations.h"
 
 // /////////////////////////////// //
 // Black Panda (STM32F4) + Harness //
