+++ board/boards/src/board_cuatro.c
@@ -1,6 +1,4 @@
-#pragma once
-
-#include "board_declarations.h"
+#include "../board_declarations.h"
 
 // ////////////////////////// //
 // Cuatro (STM32H7) + Harness //
