+++ board/boards/src/board_grey.c
@@ -1,6 +1,4 @@
-#pragma once
-
-#include "board_declarations.h"
+#include "../board_declarations.h"
 
 // //////////////////// //
 // Grey Panda (STM32F4) //
