+++ board/boards/src/board_tres.c
@@ -1,6 +1,4 @@
-#pragma once
-
-#include "board_declarations.h"
+#include "../board_declarations.h"
 
 // ///////////////////////////
 // Tres (STM32H7) + Harness //
