+++ board/boards/src/board_uno.c
@@ -1,12 +1,10 @@
-#pragma once
-
-#include "board_declarations.h"
+#include "../board_declarations.h"
 
 // /////////////////////// //
-// Dos (STM32F4) + Harness //
+// Uno (STM32F4) + Harness //
 // /////////////////////// //
 
-static void dos_enable_can_transceiver(uint8_t transceiver, bool enabled) {
+static void uno_enable_can_transceiver(uint8_t transceiver, bool enabled) {
   switch (transceiver){
     case 1U:
       set_gpio_output(GPIOC, 1, !enabled);
@@ -26,7 +24,7 @@ static void dos_enable_can_transceiver(uint8_t transceiver, bool enabled) {
   }
 }
 
-static void dos_set_led(uint8_t color, bool enabled) {
+static void uno_set_led(uint8_t color, bool enabled) {
   switch (color){
     case LED_RED:
       set_gpio_output(GPIOC, 9, !enabled);
@@ -42,13 +40,18 @@ static void dos_set_led(uint8_t color, bool enabled) {
   }
 }
 
-static void dos_set_bootkick(BootState state) {
-  set_gpio_output(GPIOC, 4, state != BOOT_BOOTKICK);
+static void uno_set_bootkick(BootState state) {
+  if (state == BOOT_BOOTKICK) {
+    set_gpio_output(GPIOB, 14, false);
+  } else {
+    // We want the pin to be floating, not forced high!
+    set_gpio_mode(GPIOB, 14, MODE_INPUT);
+  }
 }
 
-static void dos_set_can_mode(uint8_t mode) {
-  dos_enable_can_transceiver(2U, false);
-  dos_enable_can_transceiver(4U, false);
+static void uno_set_can_mode(uint8_t mode) {
+  uno_enable_can_transceiver(2U, false);
+  uno_enable_can_transceiver(4U, false);
   switch (mode) {
     case CAN_MODE_NORMAL:
     case CAN_MODE_OBD_CAN2:
@@ -60,7 +63,7 @@ static void dos_set_can_mode(uint8_t mode) {
         // B5,B6: normal CAN2 mode
         set_gpio_alternate(GPIOB, 5, GPIO_AF9_CAN2);
         set_gpio_alternate(GPIOB, 6, GPIO_AF9_CAN2);
-        dos_enable_can_transceiver(2U, true);
+        uno_enable_can_transceiver(2U, true);
       } else {
         // B5,B6: disable normal CAN2 mode
         set_gpio_mode(GPIOB, 5, MODE_INPUT);
@@ -69,7 +72,7 @@ static void dos_set_can_mode(uint8_t mode) {
         // B12,B13: OBD mode
         set_gpio_alternate(GPIOB, 12, GPIO_AF9_CAN2);
         set_gpio_alternate(GPIOB, 13, GPIO_AF9_CAN2);
-        dos_enable_can_transceiver(4U, true);
+        uno_enable_can_transceiver(4U, true);
       }
       break;
     default:
@@ -78,54 +81,65 @@ static void dos_set_can_mode(uint8_t mode) {
   }
 }
 
-static bool dos_check_ignition(void){
+static bool uno_check_ignition(void){
   // ignition is checked through harness
   return harness_check_ignition();
 }
 
-static void dos_set_ir_power(uint8_t percentage){
+static void uno_set_usb_switch(bool phone){
+  set_gpio_output(GPIOB, 3, phone);
+}
+
+static void uno_set_ir_power(uint8_t percentage){
   pwm_set(TIM4, 2, percentage);
 }
 
-static void dos_set_fan_enabled(bool enabled){
+static void uno_set_fan_enabled(bool enabled){
   set_gpio_output(GPIOA, 1, enabled);
 }
 
-static void dos_set_siren(bool enabled){
-  set_gpio_output(GPIOC, 12, enabled);
-}
-
-static bool dos_read_som_gpio (void){
-  return (get_gpio_input(GPIOC, 2) != 0);
-}
-
-static void dos_init(void) {
+static void uno_init(void) {
   common_init_gpio();
 
   // A8,A15: normal CAN3 mode
   set_gpio_alternate(GPIOA, 8, GPIO_AF11_CAN3);
   set_gpio_alternate(GPIOA, 15, GPIO_AF11_CAN3);
 
+  // GPS off
+  set_gpio_output(GPIOB, 1, 0);
+  set_gpio_output(GPIOC, 5, 0);
+  set_gpio_output(GPIOC, 12, 0);
+
   // C8: FAN PWM aka TIM3_CH3
   set_gpio_alternate(GPIOC, 8, GPIO_AF2_TIM3);
 
-  // C2: SOM GPIO used as input (fan control at boot)
-  set_gpio_mode(GPIOC, 2, MODE_INPUT);
-  set_gpio_pullup(GPIOC, 2, PULL_DOWN);
+  // Turn on phone regulator
+  set_gpio_output(GPIOB, 4, true);
 
   // Initialize IR PWM and set to 0%
   set_gpio_alternate(GPIOB, 7, GPIO_AF2_TIM4);
   pwm_init(TIM4, 2);
-  dos_set_ir_power(0U);
+  uno_set_ir_power(0U);
 
-  // Bootkick
-  dos_set_bootkick(true);
+  // Switch to phone usb mode if harness connection is powered by less than 7V
+  if(white_read_voltage_mV() < 7000U){
+    uno_set_usb_switch(true);
+  } else {
+    uno_set_usb_switch(false);
+  }
 
-  // Init clock source (camera strobe) using PWM
-  clock_source_init();
+  // Bootkick phone
+  uno_set_bootkick(BOOT_BOOTKICK);
 }
 
-static harness_configuration dos_harness_config = {
+static void uno_init_bootloader(void) {
+  // GPS off
+  set_gpio_output(GPIOB, 1, 0);
+  set_gpio_output(GPIOC, 5, 0);
+  set_gpio_output(GPIOC, 12, 0);
+}
+
+static harness_configuration uno_harness_config = {
   .has_harness = true,
   .GPIO_SBU1 = GPIOC,
   .GPIO_SBU2 = GPIOC,
@@ -139,31 +153,27 @@ static harness_configuration dos_harness_config = {
   .adc_channel_SBU2 = 13
 };
 
-board board_dos = {
-  .harness_config = &dos_harness_config,
-#ifdef ENABLE_SPI
-  .has_spi = true,
-#else
+board board_uno = {
+  .harness_config = &uno_harness_config,
   .has_spi = false,
-#endif
   .has_canfd = false,
-  .fan_max_rpm = 6500U,
+  .fan_max_rpm = 5100U,
   .fan_max_pwm = 100U,
   .avdd_mV = 3300U,
-  .fan_stall_recovery = true,
-  .fan_enable_cooldown_time = 3U,
-  .init = dos_init,
-  .init_bootloader = unused_init_bootloader,
-  .enable_can_transceiver = dos_enable_can_transceiver,
-  .set_led = dos_set_led,
-  .set_can_mode = dos_set_can_mode,
-  .check_ignition = dos_check_ignition,
+  .fan_stall_recovery = false,
+  .fan_enable_cooldown_time = 0U,
+  .init = uno_init,
+  .init_bootloader = uno_init_bootloader,
+  .enable_can_transceiver = uno_enable_can_transceiver,
+  .set_led = uno_set_led,
+  .set_can_mode = uno_set_can_mode,
+  .check_ignition = uno_check_ignition,
   .read_voltage_mV = white_read_voltage_mV,
   .read_current_mA = unused_read_current,
-  .set_fan_enabled = dos_set_fan_enabled,
-  .set_ir_power = dos_set_ir_power,
-  .set_siren = dos_set_siren,
-  .set_bootkick = dos_set_bootkick,
-  .read_som_gpio = dos_read_som_gpio,
+  .set_fan_enabled = uno_set_fan_enabled,
+  .set_ir_power = uno_set_ir_power,
+  .set_siren = unused_set_siren,
+  .set_bootkick = uno_set_bootkick,
+  .read_som_gpio = unused_read_som_gpio,
   .set_amp_enabled = unused_set_amp_enabled
 };
