+++ board/boards/unused_funcs.h
@@ -1,32 +1,17 @@
 #pragma once
 
-void unused_init_bootloader(void) {
-}
+void unused_init_bootloader(void);
 
-void unused_set_ir_power(uint8_t percentage) {
-  UNUSED(percentage);
-}
+void unused_set_ir_power(uint8_t percentage);
 
-void unused_set_fan_enabled(bool enabled) {
-  UNUSED(enabled);
-}
+void unused_set_fan_enabled(bool enabled);
 
-void unused_set_siren(bool enabled) {
-  UNUSED(enabled);
-}
+void unused_set_siren(bool enabled);
 
-uint32_t unused_read_current(void) {
-  return 0U;
-}
+uint32_t unused_read_current(void);
 
-void unused_set_bootkick(BootState state) {
-  UNUSED(state);
-}
+void unused_set_bootkick(BootState state);
 
-bool unused_read_som_gpio(void) {
-  return false;
-}
+bool unused_read_som_gpio(void);
 
-void unused_set_amp_enabled(bool enabled) {
-  UNUSED(enabled);
-}
+void unused_set_amp_enabled(bool enabled);
